Removed unused code + fixes
Removed unused code in some areas.

Fixed an issue where the parent component was getting unmounted when it shouldn't be, if a child component was updated and returned another component rather than it's previous standard dom element.

Fixed an issue where the new dom base had it's component cleared. Instead, it should have been the old dom base having it's component cleared.

Fixed an issue where the child component, if there was one, would not unmount, if the parent component returned a null value from render()